// src/interFace/interFace.ts
export interface IDayTourDataType {
  id: number;
  img: StaticImageData;
  tourTitle: string;
  tourLocation: string;
  tourTime: number;
  description: string;
}